pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Bryan Belmont", email = "bryan.belmont@seedthought.org" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langgraph>=0.6.4",
    "llama-index>=0.13.0",
    "pydantic>=2",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.5",
    "tavily-python>=0.7.10",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
ai-agents = "ai_agents:main"

[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"
